Commands base class
