+++ s4/commands/daemon_command.py
@@ -3,41 +3,11 @@ from collections import defaultdict
 
 from inotify_simple import flags
 
-from s4 import sync
-from s4.clients.local import get_local_client
-from s4.clients.s3 import get_s3_client
+from s4.commands import Command
 from s4.inotify_recursive import INotifyRecursive
 
 
-def get_clients(entry):
-    target_1 = entry['local_folder']
-    target_2 = entry['s3_uri']
-    aws_access_key_id = entry['aws_access_key_id']
-    aws_secret_access_key = entry['aws_secret_access_key']
-    region_name = entry['region_name']
-
-    # append trailing slashes to prevent incorrect prefix matching on s3
-    if not target_1.endswith('/'):
-        target_1 += '/'
-    if not target_2.endswith('/'):
-        target_2 += '/'
-
-    client_1 = get_local_client(target_1)
-    client_2 = get_s3_client(target_2, aws_access_key_id, aws_secret_access_key, region_name)
-    return client_1, client_2
-
-
-def get_sync_worker(entry):
-    client_1, client_2 = get_clients(entry)
-    return sync.SyncWorker(client_1, client_2)
-
-
-class DaemonCommand(object):
-    def __init__(self, args, config, logger):
-        self.args = args
-        self.config = config
-        self.logger = logger
-
+class DaemonCommand(Command):
     def run(self, terminator=lambda x: False):
         all_targets = list(self.config['targets'].keys())
         if not self.args.targets:
@@ -68,7 +38,7 @@ class DaemonCommand(object):
                 watch_map[wd] = target
 
             # Check for any pending changes
-            worker = get_sync_worker(entry)
+            worker = self.get_sync_worker(target)
             worker.sync(conflict_choice=self.args.conflicts)
 
         index = 0
@@ -84,8 +54,7 @@ class DaemonCommand(object):
                     to_run[target].add(event.name)
 
             for target, keys in to_run.items():
-                entry = self.config['targets'][target]
-                worker = get_sync_worker(entry)
+                worker = self.get_sync_worker(target)
 
                 # Should ideally be setting keys to sync
                 self.logger.info('Syncing {}'.format(worker))
