+++ s4/commands/ls_command.py
@@ -4,34 +4,10 @@ from datetime import datetime
 
 from tabulate import tabulate
 
-from s4.clients.local import get_local_client
-from s4.clients.s3 import get_s3_client
+from s4.commands import Command
 
 
-def get_clients(entry):
-    target_1 = entry['local_folder']
-    target_2 = entry['s3_uri']
-    aws_access_key_id = entry['aws_access_key_id']
-    aws_secret_access_key = entry['aws_secret_access_key']
-    region_name = entry['region_name']
-
-    # append trailing slashes to prevent incorrect prefix matching on s3
-    if not target_1.endswith('/'):
-        target_1 += '/'
-    if not target_2.endswith('/'):
-        target_2 += '/'
-
-    client_1 = get_local_client(target_1)
-    client_2 = get_s3_client(target_2, aws_access_key_id, aws_secret_access_key, region_name)
-    return client_1, client_2
-
-
-class LsCommand(object):
-    def __init__(self, args, config, logger):
-        self.args = args
-        self.config = config
-        self.logger = logger
-
+class LsCommand(Command):
     def run(self):
         if 'targets' not in self.config:
             self.logger.info('You have not added any targets yet')
@@ -44,7 +20,7 @@ class LsCommand(object):
             return
 
         target = self.config['targets'][self.args.target]
-        client_1, client_2 = get_clients(target)
+        client_1, client_2 = self.get_clients(target)
 
         sort_by = self.args.sort_by.lower()
         descending = self.args.descending
